requirements.txt
azure-batch>=14.2.0
azure-servicebus>=7.11.4
python-dotenv>=1.0.1
